+++ src/data.js
@@ -1,3 +1,44 @@
+//import pokemon from './data/pokemon/pokemon.js';export const searchFilter =  (pokemon,nombrePokemon)=> {
+  export const searchFilter =  (pokemon,nombrePokemon)=> {
+    const nameFilter= pokemon.filter (x=>
+      x.name.toLowerCase()==nombrePokemon.toLowerCase())
+  
+      
+    ;return nameFilter
+  
+  
+  };
+  
+
+
+/*export const searchFilter =  (pokemon,nombrePokemon)=> {
+  const nameFilter= pokemon.filter (x=>
+    x.name.toUpperCase()==nombrePokemon.toUpperCase())
+
+    
+  ;return nameFilter
+
+
+};*/
+
+
+/*export const ordenAzPokemon=(array)=>{
+
+  return array.sort(function (a, b) {
+    
+    if (a.name < b.name) {
+      return -1;
+    }
+    if (a.name> b.name) {
+      return 1;
+    }
+   return 0;
+  })
+  
+  
+}
+
+
 // estas funciones son de ejemplo
 /*export const pokemon = () => {
   return 'pokemon';
@@ -6,7 +47,7 @@ export const anotherExample = () => {
   return 'OMG';
 };*/
 
-//import pokemon from './data/pokemon/pokemon.js';
+
 //const pokemon = 
 
 /*Te recomendamos que este archivo contenga toda la funcionalidad que corresponda a obtener, procesar y manipular datos (tus funciones). Por ejemplo:
