+++ src/index.html
@@ -55,6 +55,7 @@
   <div class="searcher">
     <div>
       <input type="search" id="search" class="search comun" placeholder="search by name" autocomplete="off" spellcheck="false"/>
+    <button type ="boton" id="boton">Buscar</button>
     </div>
   <main>
     <div class="cont-card" id="cont-card"></div>
